python/requirements.txt
mysql-connector-python==8.3.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.40.0
scikit-learn==1.4.0
numpy==1.26.4